refactor: move initBackToStartButton to button_utils.js and toggle password visibility functions to input_field_utils.js
