+++ frontend/src/js/login_panel.js
@@ -1,11 +1,11 @@
 
-import { loadStartPanel } from "./start_panel.js";
 import { loadComponent } from "./utils/ui_utils.js";
 import { addEventListenerTo } from "./utils/ui_utils.js";
 import { getColor, getRandomColor } from "./utils/color_utils.js";
 import { resetInputField, setInputFieldHint } from "./utils/input_field_utils.js";
 
-let isPasswordVisible = false;
+import { initBackToStartButton } from "./utils/button_utils.js"
+import { initTogglePasswordVisibilityIcon } from "./utils/input_field_utils.js";
 
 export async function loadLoginPanel() {
   try {
@@ -18,59 +18,6 @@ export async function loadLoginPanel() {
     console.error('Error loading Login Panel:', error)
   }
 }
-
-function initBackToStartButton() {
-  const backToStartButton = document.getElementById('back-to-start-button')
-  const backIcon = document.getElementById('back-icon')
-  let colorInfo = {
-    hex: '',
-    name: ''
-  }
-
-  addEventListenerTo(
-    backToStartButton,
-    'click',
-    () => loadStartPanel()
-  )
-
-  addEventListenerTo(
-    backToStartButton,
-    'mouseover',
-    () => {
-      colorInfo = getRandomColor(500)
-
-      backToStartButton.style.backgroundColor = colorInfo['hex']
-      backIcon.style.color = getColor(colorInfo['name'], 800)
-    }
-  )
-
-  addEventListenerTo(
-    backToStartButton,
-    'mouseout',
-    () => {
-      backToStartButton.style.backgroundColor = getColor('charcoal', 700)
-      backIcon.style.color = getColor('charcoal', 100)
-    }
-  )
-
-  addEventListenerTo(
-    backToStartButton,
-    'mousedown',
-    () => {
-      backToStartButton.style.backgroundColor = getColor(colorInfo['name'], 700)
-    }
-  )
-
-  addEventListenerTo(
-    backToStartButton,
-    'mouseup',
-    () => {
-      backToStartButton.style.backgroundColor = colorInfo['hex']
-      backIcon.style.color = getColor(colorInfo['name'], 800)
-    }
-  )
-}
-
 function initConfirmLoginButton() {
   const button = document.getElementById('confirm-login-button')
   let colorInfo = {
@@ -135,37 +82,3 @@ function initConfirmLoginButton() {
     }
   )
 }
-
-function initTogglePasswordVisibilityIcon() {
-  const togglePasswordVisibilityIcon = document.getElementById('toggle-password-visibility-icon')
-
-  addEventListenerTo(
-    togglePasswordVisibilityIcon,
-    'click',
-    () => togglePasswordVisibility(togglePasswordVisibilityIcon)
-  )
-}
-
-function togglePasswordVisibility(icon) {
-  const passwordField = document.querySelector('#login-password-input-container > .password-container > .input-field')
-
-  if (!icon) {
-    console.error('Error: toggle-password-visibility-icon not found')
-    return
-  }
-  if (!passwordField) {
-    console.error('Error: login-password-input not found')
-    return
-  }
-
-  if (!isPasswordVisible) {
-    passwordField.setAttribute('type', 'text')
-    icon.innerHTML = 'visibility'
-    isPasswordVisible = true
-  } else {
-    passwordField.setAttribute('type', 'password')
-    icon.innerHTML = 'visibility_off'
-    isPasswordVisible = false
-  }
-}
-
